Append to series on Scheduled to Published
When a Scheduled Post with an unset Series Part goes from Scheduled to Published, it will be added as the last item in that Series. This means you don't need to remember which part number you are on as you schedule series post uploads.

Also fixed a perceived bug: Previously, when scheduling more that one post to be added to a Series without providing the Series Part to any of them, each subsequent scheduled post would take the "empty" Series part position, and increment all other posts in the series. As a result, if you scheduled n posts to a series, the first n-1 posts you scheduled would become the first n-1 posts in that series, incrementing all the others.